Fix serial-update-trx
